4.0 Dev Preview 3
More changes to the overall website, focusing on making the old pages
work with the new sidebar.
Social page changed to mobile version on desktops while a better
solution is worked on (probably for after the 4.0 update).
The changeling page merged with the footer for desktop, displaying more
information on that page (like last update and contact info)